refactor(code): extract opencode and pi setup from init.ts
Split the monolithic init.ts (~617 lines) into focused modules:
- opencode.ts: OpenCode config and agent setup
- pi.ts: Pi provider installation and agent setup
- utils.ts: Shared JSON file helpers (readJsonMaybe, writeJsonFile, stripJsoncComments)

init.ts is now a thin orchestrator (~158 lines) that delegates to
tool-specific modules after handling tool/scope selection and auth.

Changes:
- Replace hand-rolled pathJoin with node:path.join
- Rename internal setup* functions to init* for consistency with cmd name
- All 165 tests pass, zero test modifications required
